added authorized-by to ev-coswid-evidence-map
Added authorized-by to ev-coswid-evidence-map to that Verifiers can identify the expected provider of the reference measurements. More discussion may be needed since another reasonable approach is to add authorized-by to the triple record. This makes sense since the objective for verifier isn't to treat this as evidence, but rather to resolve the authority for who is providing the reference values.